Moved IClubInfo type to a new file
