+++ components/Widgets/BuyInWidget/BuyInWidgetWrapper.tsx
@@ -4,8 +4,8 @@ import { formatUnits } from "ethers/lib/utils";
 import { useRouter } from "next/router";
 import React, { useEffect, useState } from "react";
 import { getUsdPrice } from "../../../lib/ethereum";
+import { IClubInfo } from "../../../lib/fetchers";
 import { provider, signer } from "../../../lib/provider";
-import { IClubInfo } from "../../../pages/clubs/[id]";
 import { Button } from "../../Button/Button";
 import LoadingWidget from "../LoadingWidget";
 import DepositBuyIn from "./DepositBuyIn";
